Creating a branch to support the Nucleo F746GZ development board to help out any Amp Hour Listeners.
This board was listed in the show notes as the development board to use for examples.  While it uses the same MCU family, they are not *exactly* identical MCUs (the STM32F767ZI MCU has an extra PLL for the DSI peripheral).  The STM32F767ZI build might work on the STM32F746GZ depending on how the memory map of each peripheral is laid out.