Update CRC32.c
I am adding an edit to this PR that uses the HAL provided CRC module (although I can't tell if the CRC module is initialized in the original project). I put a block-comment copy of my CRC initilization structure at the bottom. Not sure which of these approaches is better for the repository. Thank you for taking a look at my PR!